Vastly simplify messages components
